--- assignmnet/.classpath
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<classpath>
-	<classpathentry kind="src" path="src"/>
-	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.8"/>
-	<classpathentry kind="lib" path="C:/Users/ladu/Downloads/ITC515_assignment2/lib/HttpPOST.jar"/>
-	<classpathentry kind="lib" path="C:/Users/ladu/Downloads/ITC515_assignment2/lib/jdom.jar"/>
-	<classpathentry kind="lib" path="C:/Users/ladu/Downloads/ITC515_assignment2/lib/postgresql-9.2-1000.jdbc4.jar"/>
-	<classpathentry kind="output" path="bin"/>
-</classpath>
