+++ assignmnet/bin/datamanagement/IStudent.java
@@ -1,0 +1,18 @@
+package datamanagement;
+
+public interface IStudent {
+
+    public Integer getID();
+
+    public String getFirstName();
+    public void setFirstName(String firstName);
+
+    public String getLastName();
+    public void setLastName(String lastName);
+
+    public void addUnitRecord( IStudentUnitRecord record );
+    public IStudentUnitRecord getUnitRecord( String unitCode );
+
+    public StudentUnitRecordList getUnitRecords();
+
+}
