+++ assignmnet/bin/datamanagement/IStudentLister.java
@@ -1,0 +1,11 @@
+package datamanagement;
+
+/**
+ * @author jtulip
+ */
+
+public interface IStudentLister {
+
+    public void clearStudents();
+    public void addStudent(IStudent student);
+}
