+++ assignmnet/bin/datamanagement/IUnitLister.java
@@ -1,0 +1,8 @@
+package datamanagement;
+
+public interface IUnitLister {
+
+    public void clearUnits();
+    public void addUnit(IUnit unit);
+
+}
