+++ assignmnet/bin/datamanagement/Main.java
@@ -1,0 +1,4 @@
+package datamanagement;
+
+public class Main {
+    public static void main(String[] p) {new cgCTL().execute();}    }
