+++ assignmnet/bin/datamanagement/StudentMap.java
@@ -1,0 +1,2 @@
+package datamanagement;
+public class StudentMap extends java.util.HashMap<Integer, IStudent> {}
