+++ assignmnet/bin/datamanagement/StudentUnitRecordList.java
@@ -1,0 +1,3 @@
+package datamanagement;
+
+public class StudentUnitRecordList extends java.util.ArrayList<IStudentUnitRecord> {}
