+++ assignmnet/bin/datamanagement/StudentUnitRecordMap.java
@@ -1,0 +1,2 @@
+package datamanagement;
+public class StudentUnitRecordMap extends java.util.HashMap<String, IStudentUnitRecord>{}
