+++ assignmnet/bin/datamanagement/UnitMap.java
@@ -1,0 +1,2 @@
+package datamanagement;
+public class UnitMap extends java.util.HashMap<String,IUnit> {}
