--- assignmnet/src/datamanagement/DataManagementDemo_DB.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<root>
-  <unitTable>
-    <unit uid="ITC515" name="Programming Practice" ps="50" cr="65" di="75" hd="85" ae="45" asg1wgt="20" asg2wgt="30" examwgt="50" />
-    <unit uid="ITC508" name="Systems Analysis" ps="50" cr="65" di="75" hd="85" ae="45" asg1wgt="30" asg2wgt="20" examwgt="50" />
-    <unit uid="ITC517" name="Software Engineering" ps="50" cr="65" di="75" hd="85" ae="45" asg1wgt="30" asg2wgt="30" examwgt="40" />
-  </unitTable>
-  <studentTable>
-    <student sid="1" fname="John" lname="Smith" />
-    <student sid="2" fname="Mary" lname="Contrary" />
-    <student sid="3" fname="Fred" lname="Nurke" />
-    <student sid="4" fname="Joe" lname="Bloe" />
-    <student sid="5" fname="Mark" lname="Bennett" />
-  </studentTable>
-  <studentUnitRecordTable>
-    <record sid="1" uid="ITC515" asg1="19.0" asg2="27.0" exam="50.0" />
-    <record sid="3" uid="ITC515" asg1="13.0" asg2="12.0" exam="25.0" />
-    <record sid="4" uid="ITC515" asg1="10" asg2="10" exam="24" />
-    <record sid="1" uid="ITC508" asg1="0.0" asg2="0.0" exam="25.0" />
-    <record sid="2" uid="ITC508" asg1="29.0" asg2="20.0" exam="39.0" />
-    <record sid="4" uid="ITC508" asg1="0.0" asg2="0.0" exam="48.0" />
-    <record sid="5" uid="ITC508" asg1="12.0" asg2="13.0" exam="25.0" />
-    <record sid="1" uid="ITC517" asg1="0" asg2="0" exam="0" />
-    <record sid="2" uid="ITC517" asg1="15" asg2="15" exam="25" />
-    <record sid="3" uid="ITC517" asg1="0" asg2="0" exam="0" />
-  </studentUnitRecordTable>
-</root>
-
