--- assignmnet/src/datamanagement/IStudentLister.java
@@ -1,11 +1,0 @@
-package datamanagement;
-
-/**
- * @author jtulip
- */
-
-public interface IStudentLister {
-
-    public void clearStudents();
-    public void addStudent(IStudent student);
-}
