--- assignmnet/src/datamanagement/IUnitLister.java
@@ -1,8 +1,0 @@
-package datamanagement;
-
-public interface IUnitLister {
-
-    public void clearUnits();
-    public void addUnit(IUnit unit);
-
-}
