--- assignmnet/src/datamanagement/ListUnitsCTL.java
@@ -1,10 +1,0 @@
-package datamanagement;
-public class ListUnitsCTL {
-    private UnitManager um;
-public ListUnitsCTL() {
-        um = UnitManager.UM();
-}
-            public void listUnits( IUnitLister lister ) {
-lister.clearUnits();UnitMap units = um.getUnits();
-        for (String s : units.keySet() )
-            lister.addUnit(units.get(s));}}
