--- assignmnet/src/datamanagement/Main.java
@@ -1,4 +1,0 @@
-package datamanagement;
-
-public class Main {
-    public static void main(String[] p) {new cgCTL().execute();}    }
