--- assignmnet/src/datamanagement/StudentMap.java
@@ -1,2 +1,0 @@
-package datamanagement;
-public class StudentMap extends java.util.HashMap<Integer, IStudent> {}
