--- assignmnet/src/datamanagement/StudentUnitRecordList.java
@@ -1,3 +1,0 @@
-package datamanagement;
-
-public class StudentUnitRecordList extends java.util.ArrayList<IStudentUnitRecord> {}
