--- assignmnet/src/datamanagement/StudentUnitRecordMap.java
@@ -1,2 +1,0 @@
-package datamanagement;
-public class StudentUnitRecordMap extends java.util.HashMap<String, IStudentUnitRecord>{}
