--- assignmnet/src/datamanagement/UnitMap.java
@@ -1,2 +1,0 @@
-package datamanagement;
-public class UnitMap extends java.util.HashMap<String,IUnit> {}
